fix: use float32 in test data and coerce query vector to native floats
- Test write_data now uses np.float32 to match Array(Float32) schema
- retrieve_online_documents_v2 coerces embedding to native Python floats
  before passing to \, avoiding BSON encoding errors with
  numpy float types

+++ sdk/python/tests/integration/online_store/test_mongodb_vector_search.py
@@ -133,7 +133,7 @@ def write_data() -> List[Dict[str, Any]]:
         rows.append(
             {
                 "item_id": i,
-                "embedding": list(np.random.rand(VECTOR_DIM).astype(float)),
+                "embedding": list(np.random.rand(VECTOR_DIM).astype(np.float32)),
                 "title": f"Document {i}",
             }
         )
